adding required components
