+++ lightyear/src/shared/replication/components.rs
@@ -54,28 +54,12 @@ pub struct Controlled;
 #[reflect(Component)]
 pub struct Replicating;
 
-/// Component that indicates which clients the entity should be replicated to.
-#[derive(Component, Clone, Debug, PartialEq, Reflect)]
-#[reflect(Component)]
-pub struct ReplicationTarget {
-    /// Which clients should this entity be replicated to
-    pub target: NetworkTarget,
-}
-
 /// Keeps track of the last known state of a component, so that we can compute
 /// the delta between the old and new state.
 #[derive(Component, Clone, Debug, PartialEq, Reflect)]
 #[reflect(Component)]
 pub struct Cached<C> {
     pub value: C,
-}
-
-impl Default for ReplicationTarget {
-    fn default() -> Self {
-        Self {
-            target: NetworkTarget::All,
-        }
-    }
 }
 
 /// Defines the target entity for the replication.
