+++ lightyear/src/shared/replication/hierarchy.rs
@@ -8,9 +8,10 @@ use crate::prelude::server::ControlledBy;
 use crate::prelude::{
     MainSet, NetworkRelevanceMode, PrePredicted, Replicated, Replicating, ReplicationGroup,
 };
+use crate::server::replication::send::ReplicationTarget;
 use crate::server::replication::send::SyncTarget;
 use crate::shared::replication::authority::{AuthorityPeer, HasAuthority};
-use crate::shared::replication::components::{ReplicateHierarchy, ReplicationTarget};
+use crate::shared::replication::components::ReplicateHierarchy;
 use crate::shared::replication::{ReplicationPeer, ReplicationSend};
 use crate::shared::sets::{InternalMainSet, InternalReplicationSet};
